test.html: Add dummy file to get AMI image compiling
Add a simple, blank test.html file in order to get the AMI image
to compile.

The recipe that's checked out in the AMI image refers to a test.html
file, but that is now missing, which causes the recipe to fail.
This dummy file will allow it to compile successfully.