bridgecontroller: Remove the use of QTimer
QTimer requires that the thread be running under QThread, with
a valid event loop.  We run under bare pthreads, and spend
most of the time blocking in accept() (as opposed to hanging
out in an event loop).  Therefore, if we wanted to use QTimer,
we should send an event to the main thread.

However, with the new thread pool architecture, there's no
need to defer execution of these calls (as there is a pool of
around 20 threads servicing requests).  Therefore, we simply
convert the call from QTimer into a direct call of the slot,
rather than using indirection.

A proper fix might be do post a message to the primary thread
to call our script at a later time.  This would require converting
the child threads to run under QThreads, and then orchestrate the
whole message passing architecture.  Such a change should be made
at a later time.